requirements.txt
Flask==0.11.1
Flask_Login==0.3.2
Flask_RESTful==0.3.5
Flask_Script==2.0.5
Flask_SQLAlchemy==2.1
Flask_WTF==0.12
numpy==1.11.0
pandas==0.17.1
pymssql==2.1.1
Werkzeug==0.11.10
WTForms==2.1
azure_common==1.1.4
